+++ src/mimic/src/orm/base/types/prim/mod.rs
@@ -1,9 +1,11 @@
 pub mod blob;
 pub mod decimal;
 pub mod principal;
+pub mod relation;
 pub mod ulid;
 
 pub use blob::Blob;
 pub use decimal::Decimal;
 pub use principal::Principal;
+pub use relation::Relation;
 pub use ulid::Ulid;